feat : #190 리스너 프로듀서 세팅 및 공통 message 구현
