+++ com.taken_seat.common-service/src/main/java/com/taken_seat/common_service/message/PaymentResultMessage.java
@@ -2,6 +2,8 @@ package com.taken_seat.common_service.message;
 
 import java.util.UUID;
 
+import com.taken_seat.common_service.message.enums.PaymentResultStatus;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,12 +15,8 @@ public class PaymentResultMessage {
 
 	private UUID bookingId;
 
+	private UUID paymentId;
+
 	private PaymentResultStatus status;
 
-	public enum PaymentResultStatus {
-		SUCCESS,                // 결제 성공
-		INSUFFICIENT_BALANCE,   // 잔고 부족
-		INVALID_PRICE,          // 잘못된 가격
-		COUPON_OR_MILEAGE_FAIL  // 마일리지 또는 쿠폰 사용 실패
-	}
 }
