+++ com.taken_seat.common-service/src/main/java/com/taken_seat/common_service/message/enums/PaymentResultStatus.java
@@ -1,0 +1,8 @@
+package com.taken_seat.common_service.message.enums;
+
+public enum PaymentResultStatus {
+	SUCCESS,                // 결제 성공
+	INSUFFICIENT_BALANCE,   // 잔고 부족
+	INVALID_PRICE,          // 잘못된 가격
+	COUPON_OR_MILEAGE_FAIL  // 마일리지 또는 쿠폰 사용 실패
+}
