+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/domain/model/Payment.java
@@ -38,6 +38,9 @@ public class Payment extends BaseTimeEntity {
 	private UUID bookingId;
 
 	@Column(nullable = false)
+	private UUID userId;
+
+	@Column(nullable = false)
 	private Integer price;
 
 	@Column(nullable = false)
